Changed HaloFlux Functions
This seen to resolve another outstanding issue between flow from blocks.
However there is a conflict for a coarse cell surrounded by finer block to the left and bottom... It needs to be resolved by separating which flux component is used where (i.e. the bottom cell cannot control the u-components of vectors)