Optimise gradient (#67)
# Optimising gradient and gradient-halo

There is a new test to confirm the validity of the new gradient and gradient halo the speed gain.
The new scheme is about 1.8x faster than original which should make the whole simulation 1.15x faster

This makes the model 22% to 40% faster than the current dev branch

- [x] Make New Halo script
- [x] Make new Gradient scheme
- [x] Write test for new halo scheme ~~including speed test~~
- [x] write test for new gradient scheme including speed test
- [x] Fix any bugs in refine linear and conserve elevation that might cause instability at wet dry interface
- [x] add some documentation
- [x] double check gpu vs cpu
- [x] Add test to CI
- [x] Implement wet/dry instability fix for when conserve elevation is not requested
- [x] Test wet/dry instability implementation 

# Revert prolongation from conserve elevation when cells are fully wet? 
Straight injection from Prolongation when neighbouring cell is dry affects the value of h, zs and zb. Also this action doesn't affect dzbdx. is that ok?
The changes to zb should probably reverted at the end of the step. no? 

On the other end, appling zb and h changes only to cells with wet neighbours sounds good but is a very inefficient action (50% slower!) so that should not be implemented.

- [x] Revert zb change at the end of the step for those cells affected by wet/dry 

## This seems to slow the code down a bit on the GPU (5-10%)
I wonder if we can reduce that so it does't try tp do anything on the GPU when blocks only have neighbour of the same level



* new shared mem gradient

* Started function to test new gradient function

* Fix gradien SMB issue and design test

* progress with test and SM functions

* update test

* CHange output o ftest

* optim synthread

* trying more shared

* Fix SMB

* clean-up

* Fix Gradient SM functions

* Try improvement on gradient edges

* Improve on whole gradient calculations

Creted a new Gradient function and tested it.
the new function (gradientGPUnew) is about 22% faster than the original

* further improvement for non adaptive run

* New Gradient hoping to improve mem speed

* optim occup for halo fill

* More optimisation

* Update Halo.cu

* add missing bracket

* fix missing }

* Add new functions to definitions

* Fix gradient calculation for test 0

* temporarily suspend new fill Halo GPU

* started Halo testing function

* Working on testing new Halo

* re-add the outishift and outjshift

This somehow diseapered from previous version. likely removed from a conflict

* Create NewGradient.md

* Instability Fix (#74)

* start an instability test

* Add dzbdx and dzbdy to output list

* Add/fix half step for debugging CPU

* Fix bug for refine linear

* Fix some conserve elevation issues

* Fix resolution bug in refine linear

* Add recalculatehh miror to recalczs

* Change adaptation of test to target

* Fix Prolongation in conserv Elev

* Clean up test output and enfore conserv elev

* map GPU grad ops to CPU

* Fix bugs in wall boundary

* Fix bug for gradientedge

* convert unsigned it to int

* remove unsigned

* Bug fix in Refine Linear

* Update NewGradient.md

* Update NewGradient.md

* Update NewGradient.md

* Add block images for doc

* Update NewGradient.md

* Update main.yml

* Comment redundant Bary interp

This part of the calculation is done with Refine linear so it is redundant

* Revert "Comment redundant Bary interp"

This reverts commit a86e0e4e37effdd2022b0bf219fe2e37c92b5ac1.

* Update Testing.cu

* add RevertPrologationElevation funct

* Reverting back to dhdx[halo] = T(0.0);

* Add half step for GPU

* Add zb reset on CPU

* replace initzbgrad for straight copy

* add test for zb init

* Move zb halo reset to end of function

This could make h zs and zb inconsistent for the halo in the outputs but otherwise it should be fine.

* Add Wet dry prolongation

This prevents spurious velocity where prolongationoperates. The code comes from the conserve elevation module but is applied optionally when conserve elevation is not required but the removal of spurious velocities is needed.

* Add a wet/dry restriction function

* Add Wet/dry restriction function

* Fix for GCC compiler

* change Param name for wet dry fix

* Add input keyword for wetdryfix

* fix readinput bug

* Add rainbnd to user input

* Create Wetdryfix-Conservelevation.md

* typo, phrasing on documentation

* Update NewGradient.md

* Update block_restriction.png

Co-authored-by: Alice Harang <alice.harang@niwa.co.nz>